+++ documentado.rb
@@ -1,6 +1,6 @@
-@raiz   = File.expand_path(File.join(File.dirname(__FILE__)))
-@minimo = File.join(@raiz, 'minimo')
-@heroku = File.join(@raiz, 'heroku.rb')
+@raiz        = File.expand_path(File.join(File.dirname(__FILE__)))
+@documentado = File.join(@raiz, 'documentado')
+@heroku      = File.join(@raiz, 'heroku.rb')
 
 @fecha ||= Time.now
 def fecha
@@ -17,9 +17,7 @@ remove_file "app/helpers/application_helper.rb"
 remove_file "app/views/layouts/application.html.erb"
 remove_file "app/controllers/application_controller.rb"
 
-remove_file "config/application.rb"
 remove_file "config/database.yml"
-remove_file "config/environments/production.rb"
 remove_file "config/initializers/inflections.rb"
 remove_file "config/environments/test.rb"
 remove_file "config/routes.rb"
@@ -33,15 +31,20 @@ remove_dir "test"
 remove_dir "doc"
 
 # Copia Recursivamente todos los archivos de la carpeta proyecto.
-directory @minimo, "../#{app_name}"
+directory @documentado, "../#{app_name}"
 
-# Instala Gemas: 
+# Instala Gemas
 run 'bundle install'
 
 # Creación, Migración y carga de la Base de Datos
 rake 'db:create', :env => 'development'
 rake 'db:migrate'
 rake 'cargar:usuarios'
+
+# Corre Generadores: 
+generate('simple_form:install --bootstrap')
+generate('sorcery:install --model Usuario')
+generate('squeel:initializer')
 
 # Subir aplicación a Heroku y Configurar.
 if yes? 'Desea subir aplicacion a Heroku? (y/n)'